+++ app/javascript/engine/control/select.ts
@@ -1,4 +1,4 @@
-import { Coordinate, CounterSelectionTarget, markerType } from "../../utilities/commonTypes"
+import { Coordinate, CounterSelectionTarget, MarkerType, markerType } from "../../utilities/commonTypes"
 import { hexDistance } from "../../utilities/utilities"
 import DeployAction from "../actions/DeployAction"
 import Game from "../Game"
@@ -74,6 +74,7 @@ export function selectable(map: Map, selection: CounterSelectionTarget, reselect
   if (!game) { return false }
   const target = selection.counter.unit as Unit
   if (target.isFeature && game.phase !== gamePhaseType.Deploy) { return false }
+  if (target.isMarker && target.type as MarkerType === markerType.Spotting) { return false }
   if (reselect && game.gameState) { return game.gameState.selectable(selection) }
   const same = target.playerNation === game.currentPlayerNation
   if (game.phase === gamePhaseType.Main) {