+++ app/models/utility/server.rb
@@ -2,6 +2,6 @@
 
 module Utility
   class Server
-    VERSION = "0.136"
+    VERSION = "0.137"
   end
 end